backend/requirements.txt
fastapi
uvicorn
sqlalchemy
websockets
pydantic
python-multipart
cloudinary
python-dotenv
psycopg2-binary
